Update PerfectPerspective.fx
Now the Pantomorphic mode and Asymmetrical mode has proper support for full-frame cropping (to the corners of the screen).
Also updated the comments and shader info, so that the Pantomorphic and Asymmetrical mode is not so hidden. Some users didn't know that this shader replaced PerferctPerspective2.fx, I hope it's now more clear.